+++ services/audio_suite/client/node/src/audio_suite_tempo_pitch_node.cpp
@@ -65,7 +65,8 @@ int32_t AudioSuiteTempoPitchNode::Init()
         static_cast<AudioSamplingRate>(nodeParameter_.inSampleRate)});
 
     CHECK_AND_RETURN_RET_LOG(nodeParameter_.inSampleRate != 0, ERROR, "Invalid input SampleRate");
-    nodeNeedDataDuration_  = (nodeParameter_.frameLen * MILLISECONDS_TO_MICROSECONDS) / nodeParameter_.inSampleRate;
+    nodeNeedDataDuration_ =
+        static_cast<uint64_t>(nodeParameter_.frameLen) * MILLISECONDS_TO_MICROSECONDS / nodeParameter_.inSampleRate;
     isInit_ = true;
     AUDIO_INFO_LOG("AudioSuiteTempoPitchNode::Init end");
     return SUCCESS;
