+++ services/audio_suite/client/node/src/audio_suite_voice_beautifier_node.cpp
@@ -59,7 +59,8 @@ int32_t AudioSuiteVoiceBeautifierNode::Init()
         static_cast<AudioSamplingRate>(nodeParameter_.inSampleRate)});
 
     CHECK_AND_RETURN_RET_LOG(nodeParameter_.inSampleRate != 0, ERROR, "Invalid input SampleRate");
-    nodeNeedDataDuration_  = (nodeParameter_.frameLen * MILLISECONDS_TO_MICROSECONDS) / nodeParameter_.inSampleRate;
+    nodeNeedDataDuration_ =
+        static_cast<uint64_t>(nodeParameter_.frameLen) * MILLISECONDS_TO_MICROSECONDS / nodeParameter_.inSampleRate;
     AUDIO_INFO_LOG("AudioSuiteVoiceBeautifierNode Init end");
     return SUCCESS;
 }
